palomar: split into multiple files
